requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
black==24.8.0
blinker==1.8.2
boto3==1.34.54
botocore==1.34.54
cachetools==5.4.0
certifi==2024.2.2
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.2
django-appconf==1.0.6
django-heroku==0.3.1
django-imagekit==5.0.0
django-on-heroku==1.1.2
faiss-cpu==1.8.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==21.2.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.5
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.43.0
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pilkit==3.0
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.27.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.2
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.7
watchdog==4.0.1
whitenoise==6.6.0
yarl==1.9.4
